MapMaker 1.0, released in 2006, on 18th of July
Nachtwolf's description is below

Added the mighty mapmaker version 1.0
This one is designed to work with an hex editor, I recommend XVI32.

With the mapeditor and the tileset editor, it's now possible to entirely design maps without relying on Daishiva's map editor and the pck / mcd files.

The mapeditor is a temporary solution which outputs a .map Hex string and which can be adapted to output to the lua format at a later date.